// src/app/models/ICompanyDetailsViewModel.ts
export class CompanyDetailsViewModel {
    listCustomfields: CustomManageMasterViewModel[];
    lstcustomVM: CustomValueMasterViewModel[];
    lstCustomOptions: CustomDrpChkValuesViewModel[];
    LstCompany: CompanyViewModel[];
    LstCompanyType: CompanyTypeViewModel[];
    lstcmpnyindustry: CompanyIndustryViewModel[];
    LstOwnerShipType: CompanyOwnershipViewModel[];
    AccoutnList: AccountListViewModel[];
    LstNotes: AllNoteViewModel[];
    CompanyObj: CompanyViewModel;
    customVM: CustomValueMasterViewModel;
    CustomDrpObj: CustomDrpChkValuesViewModel;
    AllNotesObj: AllNoteViewModel;
    AccountObj: AccountListViewModel;
    StateList: StateViewModel[];
    CountryList: CountryViewModel[];
}
export class CustomManageMasterViewModel {
    FieldId: number;
    UserID: number;
    Module: string;
    Column_Id: number;
    ActualColumnName: string;
    Column_Label: string;
    Column_Type: string;
    Field_Type: string;
    Column_Description: string;
    HoverText: string;
    InputDataType: string;
    DefaultValue: string;
    DisplayPosition: number | null;
    MaxLength: number | null;
    RequiredField: boolean | null;
    MultiValued: boolean | null;
    ColumnDrpChkValues: string;
    CreatedDate: Date;
    ModifiedDate: Date;
    IsActive: boolean;
    IsShowInJoinNow: boolean;
    lstCustomOptions: CustomDrpChkValuesViewModel[];
    DrpValueId: number;
    drpFieldId: number;
    DrpValue: string;
    IsDefault: boolean;
    flag: number;
    Fieldcount: number;
}

export class CustomDrpChkValuesViewModel {
    DrpValueId: number;
    FieldId: number;
    DrpValue: string;
    IsDefault: boolean;
    CreatedDate: Date | string;
    defaultval: string;
}
export class CustomValueMasterViewModel {
    CValueId: number;
    UserID: number;
    Module: string;
    ModuleRecordId: number | null;
    ModuleRecordIdTmp: number | null;
    MasterFieldId: number | null;
    CustomFieldvalue: string;
    CreatedDate: Date | null;
    ModifiedDate: Date | null;
    SearchIndex: string;
    Column_Id: number;
    ActualColumnName: string;
    Column_Label: string;
    Column_Type: string;
    Field_Type: string;
    Column_Description: string;
    HoverText: string;
    InputDataType: string;
    DefaultValue: string;
    MaxLength: number | null;
    RequiredField: boolean | null;
    lstCustomOptionsVal: CustomDrpChkValuesViewModel[];
    DrpValueId: number;
    drpFieldId: number;
    DrpValue: string;
    IsDefault: boolean;
    CMMFieldId: number;
    FieldId: number | null;
}
export class CompanyViewModel {
    ID: number;
    Name: string;
    CompanyDBA: string;
    ParentCompanyID: number | null;
    Phone: string;
    RoleName: string;
    Fax: string;
    Email: string;
    Website: string;
    CompanytypeID: number | null;
    CompanyIndustryID: number | null;
    CompanySite: string;
    CompanyCode: string;
    AnnualRevenue: string;
    CreatedDate: Date | null;
    ModifiedDate: Date | null;
    CreatedBy: number | null;
    ModifiedBy: number | null;
    TrickerSymbol: string;
    OwnerID: number | null;
    Ownership: string;
    Employees: string;
    Description: string;
    SicCode: string;
    Billingstreet: string;
    Shippingstreet: string;
    Billingcity: string;
    Shippingcity: string;
    BillingstateID: number | null;
    ShippingstateID: number | null;
    Billingzip: string;
    Shippingzip: string;
    BillingcountryID: number | null;
    ShippingcountryID: number | null;
    Naicscode: string;
    MailingAddress: string;
    BillingAddress: string;
    ReferredBy: string;
    CompanyStatusID: number | null;
    SeachIndex: string;
    ParentCompanyText: string;
    CompanyTypeText: string;
    CompanyIndustryText: string;
    CompanyOwnershipTypeText: string;
    PhoneExt: string;
    MailingStateText: string;
    MailingCountryText: string;
    BillingStateText: string;
    BillingCountryText: string;
    ParentCompanyName: string;
    MailingFullAddress: string;
    BillingFullAddress: string;
    Keyword: string;
    RoleID: number;
    ColumnName: string;
    Datatype: string;
    Maxlength: number;
    Ownership1: string;
    CompanyName: string;
    CompanyID: number;
    RecordsCount: number;
    CompanyActivityID: number;
    CompanyAccountID: number;
    IsActivity: number;
    Status: any
    Billingcounty: any;
    Mailingcounty: any
    PreferContact: any


   
  LName?: string;
  StreetAddress1?: string;
  StateOrProvinceCode?: string;
  City?: string;
  PostalCode?: string;
  CountryCode?: string;
  TotalEmployees?: string;
  MailingAddress2?: string;
  Mailingcity?: string;
  MailingstateID?: number | null;
  Mailingzip?: string;
  MailingcountryID?: number | null;
  BillingAddress2?: string;
  Company?: string;
  CustomerType?: string;
  Mobile?: string;
  Street?: string;
  State?: string;
  ZIP?: string;
  Country?: string;
  OpeningBalance?: string;
  Date?: string;
  ResaleNumber?: string;
  QBCustID?: any;
}


export class CompanyTypeViewModel {
    ID: number;
    Name: string;
    Description: string;
    Active: number | null;
    CreatedDate: Date | null;
    ModifiedDate: Date | null;
    Createdby: number | null;
    ActiveStatus: string;
    RecordsCount: number;
    DuplicateName: string;
}
export class CompanyIndustryViewModel {
    ID: number;
    Name: string;
    Description: string;
    Active: number | null;
    CreatedDate: Date | null;
    ModifiedDate: Date | null;
    Createdby: number | null;
    RecordsCount: number;
    ActiveStatus: string;
    DuplicateName: string;
}
export class CompanyOwnershipViewModel {
    ID: number;
    Name: string;
    Description: string;
    Active: number | null;
    CreatedDate: Date | null;
    ModifiedDate: Date | null;
    Createdby: number | null;
    ActiveStatus: string;
    RecordsCount: number;
    DuplicateName: string;
}
export class AccountListViewModel {

    ID: number;
    Name: string;
    Title: string;
    FirstName: string;
    MiddleName: string;
    LastName: string;
    Phone: string;
    Fax: string;
    Mobile: string;
    EmailOptOut: string;
    AccountTypeId: number | null;
    Website: string;
    Ownership: string;
    TotalEmployees: string;
    IndustryID: number | null;
    AnnualRevenue: string;
    Rating: string;
    LeadStatus: string;
    LeadSource: string;
    MailingAddress: string;
    MailingAddress2: string;
    Mailingcity: string;
    MailingstateID: number | null;
    Mailingzip: string;
    MailingcountryID: number | null;
    BillingAddress: string;
    BillingAddress2: string;
    Billingcity: string;
    BillingstateID: number | null;
    Billingzip: string;
    BillingcountryID: string;
    CreatedDate: Date | null;
    ModifiedDate: Date | null;
    CreatedBy: number | null;
    ModifiedBy: number | null;
    SeachIndex: string;
    FacebookUsername: string;
    TwitterUsername: string;
    LinkedInUsername: string;
    SkypeUsername: string;
    ContactTypeID: number | null;
    CompanyID: number | null;
    CompanyName: string;
    CompanyDBA: string;
    OwnerID: number | null;
    EstimateBudget: number | null;
    CloseDate: Date | null;
    StageID: number | null;
    BusinessTypeID: number | null;
    BusinessTypeName: string;
    Probability: string;
    NextStep: string;
    ExpectedRevenue: number | null;
    OppurtunitySourceID: number | null;
    OpportunitySource: string;
    ContactID: number | null;
    IsLead: string;
    ProdOpporID: number | null;
    GenooSource: string;
    GenooID: string;
    IsGenoo: string;
    LeadSourceText: string;
    BuyingStageText: string;
    ContactTypeText: string;
    PhoneExt: string;
    MailingStateText: string;
    MailingCountryText: string;
    BillingStateText: string;
    BillingCountryText: string;
    OptEmailOut: boolean;
    LeadName: string;
    CompanyExists: string;
    IsOpportunity: boolean;
    OppAmount: number | null;
    OppName: string;
    ContactName: string;
    Owner: string;
    OwnerFirstName: string;
    OwnerLastName: string;
    Stage: string;
    LSource: string;
    BuyingStage: string;
    MaillingCountry: string;
    MaillingStateName: string;
    MaillingAbbreviation: string;
    BillingCountry: string;
    BillingStateName: string;
    BillingAbbreviation: string;
    Keyword: string;
    RoleID: number;
    ColumnName: string;
    Datatype: string;
    Maxlength: number;
    APIService: string;
    initialMailc: string;
    ContactCompanyName: string;
    OpporContactName: string;
    NotesId: number;
    OwnerShip1: string;
    Stage1: string;
    APIContactID: string;
    IsActivity: number;
    DepartmentID: number | null;
    DepartmentText: string;
    Activestatus: string;
    IsContactActive: boolean | null;
    FilePath: string;
    Radius: number;
    ZipCode: number;
    MaillingZip: string;
    RecordID: number;
    Distance: number;
    Postal: string;
    County: string;
    City1: string;
    Lat: string;
    Long: string;
    Country: string;
    Geo: string;
    Place: string;
    RecordsCount: number;
}
export class AllNoteViewModel {
    NotesID: number;
    AccounttypeID: number | null;
    IsActivity: boolean | null;
    ContactId: number | null;
    CompanyID: number | null;
    OpportunityID: number | null;
    LeadID: number | null;
    ActivityId: number | null;
    Notes: string;
    CreatedDate: Date | null;
    ModifiedDate: Date | null;
    CreatedBy: number | null;
    ModifiedBy: number | null;
    Module: string;
    Creator: string;
    Modifier: string;
    CreDate: string;
    CreTime: string;
    Moduleid: number;
    pagesize: number;
}
export class StateViewModel {
    ID: number;
    StateName: string;
    AbbrevIation: string;
    FullStateName: string;
}
export class CountryViewModel {
    CountryId: number;
    CountryName: string;
}